Update README.rst (#7)
* Update README.rst

+++ test/src/test_readme.cpp
@@ -29,10 +29,10 @@ TEST(test_readme, reading_a_bit_field)
 {
     auto reader = bitter::reader<uint32_t, 8, 8, 8, 8>(0xdeadbeef);
 
-    uint32_t value0 = reader.field<0>().read_as<uint8_t>();
-    uint32_t value1 = reader.field<1>().read_as<uint8_t>();
-    uint32_t value2 = reader.field<2>().read_as<uint8_t>();
-    uint32_t value3 = reader.field<3>().read_as<uint8_t>();
+    uint8_t value0 = reader.field<0>().read_as<uint8_t>();
+    uint8_t value1 = reader.field<1>().read_as<uint8_t>();
+    uint8_t value2 = reader.field<2>().read_as<uint8_t>();
+    uint8_t value3 = reader.field<3>().read_as<uint8_t>();
 
     assert(value0 == 0xef); // Read bits 0-7
     assert(value1 == 0xbe); // Read bits 8-15
